\name{rca}
\alias{rca}
\title{Relevant Component Analysis}
\usage{
  rca(df)
}
\arguments{
  \item{df}{data frame to check}
  \item{df2}{data frame to check 2}
}
\description{
  Performs relevant components analysis (RCA) on the given data matrix
  and returns the results.
}
\keyword{rca,relevant,component}

