% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdml-package.R
\docType{package}
\name{sdml-package}
\alias{sdml}
\alias{sdml-package}
\title{Supervised Distance Metric Learning with R}
\description{
The sdml package aims to implement the state-of-the-art
algorithms for supervised distance metric learning.
These methods are widely applied in clustering,
classification, information retrieval,
and computer vision problems.
}
\details{
\tabular{ll}{ Package: \tab sdml\cr Type: \tab Package\cr
Version: \tab 0.1\cr License: \tab BSD\cr } Additional Desc.
}
\note{
Bug reports and feature requests could be sent to
\url{https://github.com/road2stat/sdml/issues}.
}
\section{Relevant Component Analysis}{
 Put RCA desc. here.
}

\section{Local Fisher Discriminant Analysis}{
 Put LFDA desc. here.
}
\examples{

}
\author{
Xiao Nan <\url{http://www.road2stat.com}>
        Gao Tao <\url{http://www.gaotao.name}>
}
\references{
The associated website for this package:
\url{https://github.com/road2stat/sdml/}
}
\seealso{
Add \code{\link{rca}}, \code{\link{lfda}},
             \code{\link{klfda}}, \code{\link{dca}},
             \code{\link{kdca}}, etc.
}
\keyword{dimension}
\keyword{distance}
\keyword{extraction}
\keyword{feature}
\keyword{learning}
\keyword{metric}
\keyword{package}
\keyword{reduction}
\keyword{sdml}
\keyword{supervised}

