% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdml-package.R
\docType{package}
\name{sdml-package}
\alias{sdml}
\alias{sdml-package}
\title{sdml: Supervised Distance Metric Learning with R}
\description{
The sdml package aims to implement the state-of-the-art
    algorithms for supervised distance metric learning.
    It includes global and local methods such as
    (Kernel) Relevant Component Analysis,
    (Kernel) Discriminative Component Analysis,
    (Kernel) Local Fisher Discriminant Analysis, etc.
    These distance metric learning methods are widely applied in
    feature extraction, dimensionality reduction, clustering,
    classification, information retrieval,
    and computer vision problems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nanxstats/sdml}
  \item Report bugs at \url{https://github.com/nanxstats/sdml/issues}
}

}
\keyword{internal}
